# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Simple TODO  API Documentation",
        "description": "Service Documentation",
        "contact": {
            "email": "info@std.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/todo/v1",
            "description": "STD API Server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authority: Any | Create a new user",
                "description": "Creates a new user account",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUser"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User account created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authority: Any | User Login",
                "description": "Authenticate an existing user and retrieve a token.",
                "operationId": "e7e8613e60681ad04c87a47062e1ee16",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "Your email address",
                                    "password": "Your password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Authentication. Access Token Returned",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": "true",
                                    "status_code": 200,
                                    "token_type": "Bearer",
                                    "expires_in": 1296000,
                                    "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJyJhdWQiOiIyIiwianRpIjoiNWJmNDY4OTFhYmJmYjk1YzQwN2E1MzNkZjMxNmEzNDMzZDBlZmI2ZDc4NmI0NmMwNjgwODFjODg0N2E3MDVjYTk4ZDNiMTVlODI1ZDcwYjQiLCJpYXQiaX0...."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/email/verification/resend": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authority: Any | Resend Verification link",
                "operationId": "82de378512bf3903c7154d2f8fb5e7d4",
                "responses": {
                    "200": {
                        "description": "Verification link is sent"
                    },
                    "401": {
                        "description": "Unauthenticated. Token Needed."
                    },
                    "403": {
                        "description": "Unauthorized. User not with access role"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authority: Any | Logout the currently logged in user",
                "description": "Revokes token",
                "operationId": "87c8bcc7029e5c36c30e15f947746d68",
                "responses": {
                    "200": {
                        "description": "User Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthenticated. Token needed"
                    }
                }
            }
        },
...
        "/update-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Resets password",
                "description": "This endpoint resets a user's password",
                "operationId": "UpdatePassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "token": "ClRIsugfJCVLuQSnr",
                                    "password": "password",
                                    "password_confirmation": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated. Token Needed."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "The name of the User",
                        "type": "string",
                        "example": "Peter Oduntan"
                    },
                    "email": {
                        "title": "email",
                        "description": "The email of the User",
                        "type": "string",
                        "example": "peter_oduntan@yahoo.com"
                    },
                    "thumbnail": {
                        "title": "thumbnail",
                        "description": "The image of the User",
                        "type": "string",
                        "example": "https://example.com/image.png"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-09-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-09-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
            "CreateUser": {
                "title": "User Creation Request",
                "description": "User Request Body Data",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "The name of the user",
                        "type": "string",
                        "example": "Peter"
                    },
                    "email": {
                        "title": "email",
                        "description": "The email of the user",
                        "type": "string",
                        "example": "peter_oduntan@yahoo.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "The password of the user",
                        "type": "string",
                        "example": "Secret123"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}